+++ templates/_eslintrc_javascript.json.ts
@@ -5,7 +5,8 @@ const templateFunction: TemplateFunction = answers => {
 
 	// This version is intended for use in JS projects
 	if (answers.language !== "JavaScript") return;
-	const useReact = answers.adminReact === "yes";
+	const useReact =
+		answers.adminReact === "yes" || answers.tabReact === "yes";
 
 	const useESLint = answers.tools && answers.tools.indexOf("ESLint") > -1;
 	if (!useESLint) return;