+++ templates/admin/tab_m.html.ts
@@ -5,11 +5,15 @@ export = (answers => {
 	const supportTab = answers.adminFeatures && answers.adminFeatures.indexOf("tab") > -1;
 	if (!supportTab) return;
 
+	const useReact = answers.tabReact === "yes";
+
 	const template = `
 <html>
 
 <head>
-	<link rel="stylesheet" type="text/css" href="../../lib/css/materialize.css">
+${useReact ? 
+`	<script type="text/javascript" src="../../socket.io/socket.io.js"></script>` :
+`	<link rel="stylesheet" type="text/css" href="../../lib/css/materialize.css">
 	<link rel="stylesheet" type="text/css" href="../../css/adapter.css" />
 
 	<script type="text/javascript" src="../../lib/js/jquery-3.2.1.min.js"></script>
@@ -25,11 +29,18 @@ export = (answers => {
 		.m span{
 			font-size: 0.9em;
 		}
-	</style>
+	</style>`}
 </head>
 
 <body>
 
+${useReact ? `
+	<!-- this is where the React components are loaded into -->
+	<div class="m adapter-container" id="root"></div>
+
+	<!-- load compiled React scripts -->
+	<script type="text/javascript" src="build/tab.js"></script>
+` : `
 	<div class="m adapter-container">
 		<div class="row">
 			<!-- Forms are the standard way to receive user inputted data.
@@ -81,7 +92,7 @@ export = (answers => {
 			</div>
 		</div>
 	</div>
-
+`}
 </body>
 
 </html>